complete basic styling for Resources
