+++ src/layouts/Resources.jsx
@@ -4,13 +4,55 @@ import axios from 'axios';
 import PropTypes from 'prop-types';
 import * as Bluebird from 'bluebird';
 import * as _ from 'lodash';
-import { prettifyResourceFileName } from '../utils';
-import ResourceCard from '../components/ResourceCard';
+import { prettifyResourceFileName, prettifyResourceCategory } from '../utils';
+import ResourceSettingsModal from '../components/ResourceSettingsModal';
 import ResourceCategoryModal from '../components/ResourceCategoryModal';
 import Header from '../components/Header';
 import Sidebar from '../components/Sidebar';
 import elementStyles from '../styles/isomer-cms/Elements.module.scss';
 import contentStyles from '../styles/isomer-cms/pages/Content.module.scss';
+
+// Constants
+const RADIX_PARSE_INT = 10;
+
+const ResourceCard = ({
+  fileName, siteName, category, settingsToggle, resourceIndex,
+}) => {
+  const { title, date, type } = prettifyResourceFileName(fileName);
+  return (
+    <div className={`${contentStyles.resource} ${contentStyles.card} ${elementStyles.card}`}>
+      <Link to={`/sites/${siteName}/resources/${category}/${fileName}`}>
+        <div id={resourceIndex} className={contentStyles.resourceInfo}>
+          <div className={contentStyles.resourceCategory}>{prettifyResourceCategory(category)}</div>
+          <h1 className={contentStyles.resourceTitle}>{title}</h1>
+          <p className={contentStyles.resourceDate}>{date}</p>
+          <p className={contentStyles.resourceType}>{type}</p>
+          <p className={contentStyles.resourceUpdated}>Updated 2 days ago</p>
+        </div>
+      </Link>
+      <button
+        type="button"
+        id={`settings-${resourceIndex}`}
+        onClick={settingsToggle}
+        className={contentStyles.resourceIcon}
+      >
+        <i id={`settingsIcon-${resourceIndex}`} className="bx bx-cog" />
+      </button>
+    </div>
+  );
+};
+
+const CreateResourceCard = ({ settingsToggle }) => (
+  <button
+    type="button"
+    id="settings-NEW"
+    onClick={settingsToggle}
+    className={`${elementStyles.card} ${contentStyles.card} ${elementStyles.addNew}`}
+  >
+    <i id="settingsIcon-NEW" className={`bx bx-plus-circle ${elementStyles.bxPlusCircle}`} />
+    <h2 id="settingsText-NEW">Add a new resource</h2>
+  </button>
+);
 
 export default class Resources extends Component {
   constructor(props) {
@@ -18,8 +60,13 @@ export default class Resources extends Component {
     this.state = {
       resourceCategories: [],
       resourcePages: [],
-      newPageName: null,
       categoryModalIsActive: false,
+      settingsIsActive: false,
+      selectedResourcePage: {
+        isNewPost: true,
+        category: '',
+        fileName: '',
+      },
     };
   }
 
@@ -65,20 +112,50 @@ export default class Resources extends Component {
     }
   }
 
-  updateNewPageName = (event) => {
-    event.preventDefault();
-    this.setState({ newPageName: event.target.value });
-  }
-
   categoryModalToggle = () => {
     this.setState((currState) => ({
       categoryModalIsActive: !currState.categoryModalIsActive,
     }));
   }
 
+  settingsToggle = (event) => {
+    const { id } = event.target;
+    const idArray = id.split('-');
+    let resourcePageIndex = idArray[1];
+
+    // User clicked on the "Create New Resource" button
+    if (resourcePageIndex === 'NEW') {
+      this.setState({
+        settingsIsActive: true,
+        selectedResourcePage: {
+          isNewPost: true,
+        },
+      });
+    } else if (resourcePageIndex === 'CLOSE') {
+      // User clicked on the close button in the settings modal
+      this.setState({ settingsIsActive: false });
+    } else {
+      // User clicked on the settings icon on an existing resource
+      resourcePageIndex = parseInt(resourcePageIndex, RADIX_PARSE_INT);
+
+      this.setState((currState) => ({
+        settingsIsActive: true,
+        selectedResourcePage: {
+          isNewPost: false,
+          category: currState.resourcePages[resourcePageIndex].category,
+          fileName: currState.resourcePages[resourcePageIndex].fileName,
+        },
+      }));
+    }
+  }
+
   render() {
     const {
-      resourceCategories, resourcePages, newPageName, categoryModalIsActive,
+      resourceCategories,
+      resourcePages,
+      categoryModalIsActive,
+      settingsIsActive,
+      selectedResourcePage,
     } = this.state;
     const { match, location } = this.props;
     const { siteName } = match.params;
@@ -88,6 +165,17 @@ export default class Resources extends Component {
         {/* main bottom section */}
         <div className={elementStyles.wrapper}>
           <Sidebar siteName={siteName} currPath={location.pathname} />
+          { settingsIsActive
+            ? (
+              <ResourceSettingsModal
+                siteName={siteName}
+                isNewPost={selectedResourcePage.isNewPost}
+                category={selectedResourcePage.category}
+                fileName={selectedResourcePage.fileName}
+                settingsToggle={this.settingsToggle}
+              />
+            )
+            : null}
 
           {/* main section starts here */}
           <div className={contentStyles.mainSection}>
@@ -102,51 +190,25 @@ export default class Resources extends Component {
             </div>
 
             <div className={contentStyles.contentContainerBoxes}>
-            <div className={contentStyles.boxesContainer}>
-              {/* Display resource cards */}
-              {resourcePages.length > 0
-                ? (
-                  <>
-                    {resourcePages.map((resourcePage) => (
-
-                      <div className={`${contentStyles.resource} ${contentStyles.card} ${elementStyles.card}`}>
-                        <a href={`/sites/${siteName}/resources/${resourcePage.category}/${resourcePage.fileName}`}>
-                        <div className={contentStyles.resourceInfo}>
-                          <div className={contentStyles.resourceCategory}>Press Releases</div>
-                          <h1 className={contentStyles.resourceTitle}>Resource Title</h1>
-                          <p className={contentStyles.resourceDate}>20 Oct 2019</p>
-                          <p className={contentStyles.resourceType}>Post</p>
-                          <p className={contentStyles.resourceUpdated}>Updated 2 days ago</p>
-                        </div>
-                        <button type="button" onClick={this.settingsToggle} className={contentStyles.resourceIcon}>
-                          <i className='bx bx-cog'></i>
-                        </button>
-                        </a>
-                      </div>
-                      // <ResourceCard
-                      //   type={resourcePage.type}
-                      //   category={resourcePage.category}
-                      //   title={resourcePage.title}
-                      //   date={resourcePage.date}
-                      //   fileName={resourcePage.fileName}
-                      //   siteName={siteName}
-                      //   resourceCategories={resourceCategories}
-                      //   isNewPost={false}
-                      // />
-                    ))}
-                  </>
-                )
-                : 'Resources loading...'}
-              <div className={`${elementStyles.card} ${contentStyles.card} ${elementStyles.addNew}`}>
-                <i className={`bx bx-plus-circle ${elementStyles.bxPlusCircle}`}></i>
-                <h2>Add a new resource</h2>
+              <div className={contentStyles.boxesContainer}>
+                {/* Display resource cards */}
+                {resourcePages.length > 0
+                  ? (
+                    <>
+                      {resourcePages.map((resourcePage, resourceIndex) => (
+                        <ResourceCard
+                          category={resourcePage.category}
+                          fileName={resourcePage.fileName}
+                          siteName={siteName}
+                          settingsToggle={this.settingsToggle}
+                          resourceIndex={resourceIndex}
+                        />
+                      ))}
+                      <CreateResourceCard settingsToggle={this.settingsToggle} />
+                    </>
+                  )
+                  : 'Loading Resources...'}
               </div>
-              <ResourceCard
-                siteName={siteName}
-                resourceCategories={resourceCategories}
-                isNewPost
-              />
-            </div>
             </div>
           </div>
           {/* main section ends here */}
@@ -165,4 +227,16 @@ Resources.propTypes = {
   location: PropTypes.shape({
     pathname: PropTypes.string.isRequired,
   }).isRequired,
+};
+
+ResourceCard.propTypes = {
+  fileName: PropTypes.string.isRequired,
+  siteName: PropTypes.string.isRequired,
+  category: PropTypes.string.isRequired,
+  settingsToggle: PropTypes.func.isRequired,
+  resourceIndex: PropTypes.number.isRequired,
+};
+
+CreateResourceCard.propTypes = {
+  settingsToggle: PropTypes.func.isRequired,
 };
